use functools
