+++ magicclass/functools/__init__.py
@@ -1,0 +1,12 @@
+"""
+A magic-class submodule that mimics the built-in ``functools`` module.
+"""
+
+from ._partial import partial, partialmethod
+from ._wraps import wraps
+
+
+__all__ = [
+    "partial",
+    "partialmethod",
+]
