+++ magicclass/functools/_partial.py
